requirements/base.txt
django==5.0.4
django-jazzmin==2.6.1
django-rosetta==0.10.0

python-environ==0.4.54

aiogram==3.4.1

loguru==0.7.2
uvloop==0.19.0
